Move conversion to converter class similar to Oculus
