+++ XRTK.WindowsMixedReality/Packages/com.xrtk.wmr/Controllers/Hands.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 64704bfeb12e32841990ffa095b4a010
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
