+++ XRTK.WindowsMixedReality/Packages/com.xrtk.wmr/Controllers/Hands/WindowsMixedRealityHandControllerDataProvider.cs.meta
