+++ XRTK.WindowsMixedReality/Packages/com.xrtk.wmr/Controllers/Hands/WindowsMixedRealityHandDataConverter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 25d284a3f798cfa4bbe5a4c86eb9527b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {fileID: 2800000, guid: 8ac5213854cf4dbabd140decf8df1946, type: 3}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
